server2/requirements.txt
flask
flask-socketio
eventlet
python-dotenv
supabase
